Logo: embedded the characters that we need to write OpenCOR.
This is so that we can see our SVG file fine on Windows and Linux since they don't have the font we use.

So, we ended up:
- editing the TTF file for the "Brush Script MT" font, so that it would only contain the characters 'C', 'O', 'R', 'e', 'n', and 'p' (done using https://fontforge.org/);
- converting the resulting TTF file to base64 (done using https://www.giftofspeed.com/base64-encoder/); and
- embedding the resulting base64 string to our SVG file (using the approach described at https://vecta.io/blog/how-to-use-fonts-in-svg#:~:text=The%20hard%20way).

Also removed the metadata that was in our SVG file.

Otherwise, "OpenCOR" used to be oblique in our SVG file, but it doesn't show oblique anymore (!?). This somehow got lost in the process (??). So, we replaced "oblique" with "normal" in our SVG file. In the end, it's better this way since "OpenCOR" in the .ico, .icns, and .png version of our logo wasn't oblique anyway! (I had just never noticed! Then again, it doesn't matter since we don't use our SVG file directly in OpenCOR -- while we do in our Web App.)