Don't need a Vite plugin for .node files.
Our issue was that we were effectively trying to access our .node file from the renderer and this required using the require() function which we simply can't use from a Web browser. So, instead, we need to access our .node file from node.js and then use contextBridge so that we can get access to our .node file from a Web browser.