Replaced rimraf with our own cleaning script.
Indeed, we want the script to fail silently if a file/directory we want to delete doesn't exist. Not only that, but we also want to be able to clean things even if no dependencies are installed while if we were to do that we were told that rimraf couldn't be found (and rightly so).