Splash screen: reworked the way configure our splash screen.
Indeed, it used to work fine when package.json didn't have `type` set to `module`, but now that we do we are not able to retrieve the copyright and version information on Windows (all fine on macOS). I tried various things, but to no avail and not to have `type` set to `module` is not an option (since CommonJS is on its way out and that we should therefore use ES modules). So, now, we configure our splash screen HTML file using CMake, resulting in a static splash screen, which means that it works on all of our supported platforms.