GHA: use a specific version of Windows, Ubuntu, and macOS.
Until now, we were using `windows-latest`, `ubuntu-latest`, and `macos-latest`, but this is risky. Indeed, GitHub sometimes (and rightly so) update the target version of `xxx-latest` and this is exactly what they just did by having `ubuntu-latest` point to `ubuntu-24.04` rather than `ubuntu-22.04` and it looks like it might be causing us some issues (not least since some jobs were run using `ubuntu-22.04` and others using `ubuntu-24.04`!). So, now, we use the specific version of Windows, Ubuntu, and macOS that we want to use.