NAPI: revisited the way we access libOpenCOR.
We originally wanted to use the same API for both the C++ and the WASM version of libOpenCOR (hence the interface), but this was not going to work in the long term. For instance, to manage a file requires different function signatures in C++ and in WASM. So, rather than having a LOCAPI interface, we now have a LOCAPI class which methods can handle both the C++ and WASM APIs.