Vue: mark some objects raw.
Some objects were being converted to a proxy which means that they couldn't be used directly in some cases. The only way to use them was to convert them back to a raw object using toRaw(). This was always done after finding out that something wasn't working in the first place.

In the process, we also removed some calls to toRaw() which were not needed (anymore?).