CD: don't do any Git check when trying to publish on npm.
Indeed, `pnpm` is unable to determine the current Git branch because we are in a detached HEAD state. One solution would be to have:
```
- name: Check out OpenCOR
  uses: actions/checkout@v4
  with:
    fetch-depth: 0
    ref: main
```
or use `--no-git-checks` like we have done here and that will, in the future, allow to manually publish to npm from any branch, should that ever be needed.