CD: reverting to `NODE_AUTH_TOKEN` when publishing to npm.
For some reasons, I can't seem to be able to publish to npm using trusted publishing. It all works fine in libOpenCOR, so not sure why it's not working here. Could it be because on npm, the package is called `opencor` (under the `@opencor` organisation) while on GitHub the repo is called `webapp`? For libOpenCOR, it's `libopencor` both on npm and GitHub.