Use of our own CORS proxy handler for Cloudflare Workers.
Until now, we were using corsproxy.io, but today it was dead for several hours and... well, we clearly don't want that. So, we came up with a simple CORS proxy handler for Cloudflare Workers.

The nice thing is that the free version of Cloudflare Workers (see https://developers.cloudflare.com/workers/platform/pricing/#workers) comes with 100,000 requests per day, no charge for duration, and 10 ms of CPU time per invocation, which is way more than we need.