Simulation Experiment UI view: show/hide the issues/graph panels.
To use `v-if`/`v-else` was completely wrong since a simulation may or not run successfully. Not only that, but this was causing some problems when issues were being shown and when a simulation was then successful: the graph panels wouldn't show unless we were to run another successful simulation. This is because the graph panels would be (re)mounted and `GraphPanelWidget` doesn't handle `vue.onMounted()` (to do what it's doing in `vue.watch()`. We could handle `vue.onMounted()`, but the better approach is indeed to use `v-show` rather than `v-if`/`v-else`.