Ensure that our splash screen is always shown.
Before, we could have had a race condition resulting in the splash screen sometimes being shown and sometimes not.

Here, we create the main window only once we know for sure that the splash screen has been created. Not also that we "manually" show the splash screen ourselves rather than rely on the `show` property being `true` by default to do it for us. This avoids potentially getting a temporary white window.