Make sure that the desktop version of OpenCOR can connect to GitHub.
The issue is only with `bun start` and therefore a packaged version of OpenCOR since `bun dev` works fine without those changes. It's just that without those changes, we can't connect to GitHub and we get the following error message from Firebase: "This operation is not supported in the environment this application is running on. "location.protocol" must be http, https or chrome-extension and web storage must be enabled. (auth/operation-not-supported-in-this-environment)."