package.json: removed `mathjs` and `plotly.js-gl2d-dist-min` from our peer dependencies.
They were originally here so that we could easily find out when a new version of them is available (since we "manually" import them), but it's not good to have things under `peerDependencies` unless actually needed. So, removing them... which means that we will need to "manually" check whether a new version of `mathjs` and/or `plotly.js-gl2d-dist-min` is available.