locCommon: try to fetch a file through our CORS proxy and if it fails then try directly.
Indeed, if we were trying to download a file from http://localhost, OpenCOR would first try to download it directly and that would fail because of CORS which would result in a message in the console. Then, it would try to download the file through OpenCOR's CORS proxy and it would be fine. So, better to try through OpenCOR's CORS proxy first and then directly if needed.